Split up navigation extensions into descriptive files
